requirements.txt
streamlit
groq
pdfplumber
pydantic
sentence-transformers
scikit-learn
python-dotenv
rapidfuzz
pandas
openpyxl
